Moving hook implementations to their own helper
